Extract /validate-imagemaid, /autosave-imagemaid, /start-imagemaid, /stop-imagemaid, /imagemaid-status into blueprints/imagemaid_routes.py (PR F)
Moves the five ImageMaid runtime routes plus the
_schedule_quickstart_imagemaid_run_marker helper out of quickstart.py.

## What moved

Five routes:
* POST /validate-imagemaid     -> validate_imagemaid     (15 lines)
* POST /autosave-imagemaid     -> autosave_imagemaid     (23 lines)
* POST /start-imagemaid        -> start_imagemaid        (93 lines)
* POST /stop-imagemaid         -> stop_imagemaid         (72 lines)
* GET  /imagemaid-status       -> imagemaid_status       (179 lines)

Plus:
* _schedule_quickstart_imagemaid_run_marker (37 lines) -- the helper
  that arms a background log-watcher thread after launching ImageMaid,
  called from modules/process_control.py via quickstart._schedule_...
  and patched by tests via qs_module._schedule_...  Kept reachable as
  a re-export.

Plus a single constant:
* IMAGEMAID_STARTUP_GRACE_SECONDS = 10 -- the grace period during which
  a still-spawning ImageMaid is reported as 'starting' instead of
  'not started'.

## Naming

blueprints/imagemaid_updates.py already exists for the install/probe/
update flow (e.g. /probe-imagemaid-root, /update-imagemaid).  This new
module owns the **runtime** flow.  They share no routes -- the split is
about purpose, not artificial line-count splitting.

## Test-patch boundary -- and why every route opens with `_qs` aliases

The existing test suite (test_start_stop.py alone has 158
monkeypatch.setattr(qs_module, ...) calls) patches helpers on the
quickstart module object.  Those patches do NOT propagate to a
blueprint that imports the helpers directly from
modules.imagemaid / modules.process_control.

To keep tests untouched, every route body now opens with::

    import quickstart as _qs
    _validate_imagemaid_settings = _qs._validate_imagemaid_settings
    _persist_imagemaid_validation = _qs._persist_imagemaid_validation
    ...

so that every leading-underscore helper resolves through the
quickstart module's namespace -- where the monkeypatches live.

The rest of the route body is **byte-for-byte verbatim** from develop
(only the @app.route -> @bp.route decorator was rewritten).
AST-verified -- all 6 moved functions match the develop source after
stripping the _qs preamble.

This is the same pattern PR E (#1324) used for the bundle helpers in
import_config_preview, and matches the convention already followed by
modules/process_control.py::resolve_maintenance_window_live, which
dispatches back to quickstart._get_maintenance_window_live so that
test patches are honored.

## Cross-module dependencies

Two helpers still live in quickstart.py and are reached via the same
_qs lazy mechanism:

* _get_active_work_blocker -- depends on the in-module
  ACTIVE_WORK_POLICIES dict + background-job registry.  Will move when
  the background-job glue is consolidated (out of scope here).
* _ingest_completed_live_logs -- depends on the live-log tailer state.
  Will move with the live-log subsystem.

## Tests

* 791 tests pass (zero regressions).  After the FIRST attempt -- where
  I made the blueprint resolve helpers via direct module imports --
  8 test_start_stop.py tests failed because monkeypatch.setattr on
  qs_module didn't reach the blueprint's bindings.  The _qs preamble
  pattern was the fix.
* ruff check clean on both touched files.
* All 10 imagemaid routes register correctly (5 from the new blueprint
  + 5 from the existing imagemaid_updates blueprint).
* AST-verified verbatim extraction: all 6 moved functions are
  byte-for-byte identical to develop after stripping the documented
  _qs preamble.

## Size impact

* quickstart.py:                       7,397 -> 6,976 lines (-421, -5.7%)
* blueprints/imagemaid_routes.py:                564 lines (new, under 600)
* tests/conftest.py:                            +14 lines (imagemaid_routes_module fixture)

## Cumulative refactor progress

Starting baseline (post #1320):    11,087 lines in quickstart.py.
After PR A (#1321):                 9,831  (-1,256)
After PR C (#1322):                 9,227  (-1,860)
After PR D (#1323):                 8,709  (-2,378)
After PR E (#1324):                 7,397  (-3,690)
**After PR F (this):                6,976  (-4,111)**

37.1% of the original quickstart.py is gone.  Audit target is ~5,000
lines (54% reduction from baseline) -- we are well over two-thirds of
the way there.
